requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.3.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39hd77b12b_8
ca-certificates=2024.11.26=haa95532_0
certifi=2024.12.14=py39haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
craft-text-detector=0.4.3=pypi_0
cuda-cccl=12.4.127=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.4.127=0
cuda-runtime=11.8.0=0
datasets=1.12.1=pypi_0
dill=0.3.9=pypi_0
filelock=3.13.1=py39haa95532_0
freetype=2.12.1=ha860e81_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gdown=5.2.0=pypi_0
giflib=5.2.2=h7edc060_0
gmpy2=2.1.2=py39h7f96b67_0
huggingface-hub=0.27.0=pypi_0
idna=3.7=py39haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.4=py39haa95532_1
jiwer=3.0.5=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.22=h5bf469e_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=h44ae7cf_1
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py39h2bbff1b_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.11=py39h827c3e9_0
mkl_random=1.2.8=py39hc64d2fc_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
multidict=6.1.0=pypi_0
multiprocess=0.70.17=pypi_0
networkx=3.2.1=py39haa95532_0
numpy=1.22.4=pypi_0
opencv-python=4.5.4.60=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=py39h096bfcc_1
pip=24.2=py39haa95532_0
propcache=0.2.1=pypi_0
pyarrow=18.1.0=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.21=h8205438_1
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.5.1=py3.9_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pyyaml=6.0.2=py39h827c3e9_0
rapidfuzz=3.11.0=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=py39haa95532_1
sacremoses=0.1.1=pypi_0
safetensors=0.5.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.1.0=py39haa95532_0
six=1.17.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tokenizers=0.21.0=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.47.1=pypi_0
typing_extensions=4.12.2=py39haa95532_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=py39haa95532_0
vc=14.40=haa95532_2
vs2015_runtime=14.42.34433=h9531ae6_2
wheel=0.44.0=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.18.3=pypi_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
